Moves HTLC related scripts to htlc directory
