+++ src/offchain/package.json
@@ -7,10 +7,10 @@
   "scripts": {
     "build": "tsc",
     "dev": "tsx watch index.ts",
-    "lock": "tsx lock.ts",
-    "refund": "tsx refund.ts",
+    "htlc:lock": "tsx htlc/lock.ts",
+    "htlc:refund": "tsx htlcrefund.ts",
     "start": "tsx index.ts",
-    "claim": "tsx claim.ts",
+    "htlc:claim": "tsx htlc/claim.ts",
     "vesting:lock": "tsx vesting/lock.ts",
     "vesting:claim": "tsx vesting/claim.ts",
     "test": "echo \"Error: no test specified\" && exit 1",
